Display shred version in solana-genesis
